Extract CLI module
