+++ bin/htpasswd-ruby
@@ -16,4 +16,4 @@ rescue LoadError
   retry
 end
 
-HTAuth::Passwd.new.run(ARGV, ENV)
+HTAuth::CLI::Passwd.new.run(ARGV, ENV)
