+++ lib/htauth.rb
@@ -31,16 +31,15 @@ end
 require 'htauth/version'
 require 'htauth/algorithm'
 require 'htauth/crypt'
-require 'htauth/digest'
 require 'htauth/digest_entry'
 require 'htauth/digest_file'
 require 'htauth/entry'
 require 'htauth/errors'
 require 'htauth/file'
 require 'htauth/md5'
-require 'htauth/passwd'
 require 'htauth/passwd_entry'
 require 'htauth/passwd_file'
 require 'htauth/plaintext'
 require 'htauth/sha1'
-
+require 'htauth/cli/digest'
+require 'htauth/cli/passwd'
